appyters/CFDE-C2M2-FAIR-Datapackage-Assessment/requirements.txt
datapackage
deriva-datapackage @ git+https://github.com/nih-cfde/deriva-datapackage@v1.0.5
fsspec
matplotlib
pandas
pyyaml
requests
sqlalchemy